Renew of refactor for Card and Exchange parts
