+++ src/main/java/Card/CardController.java
@@ -4,24 +4,19 @@ import App.ControllerInterface;
 import DataObjects.UserObject;
 import LogIn.LoggedIn.LoggedInController;
 import LogIn.Welcome.WelcomeController;
-import org.jetbrains.annotations.NotNull;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 public class CardController implements ControllerInterface {
-    public static List<Card> cardList = new ArrayList<>();
+    public static List<CardObject> cardList = new ArrayList<>();
     static UserObject loggedInUser;
     private CardPresenter cardPresenter;
-    private WelcomeController welcomeController;
     static CardDBAccess cardDBAccess = new CardDBAccess();
 
     public CardController(UserObject user) {
         this.loggedInUser = user;
         this.cardPresenter = new CardPresenter(this);
-        this.welcomeController = new WelcomeController();
     }
 
     @Override
@@ -35,30 +30,19 @@ public class CardController implements ControllerInterface {
         controller.launch();
     }
 
-    /**
-     * used to load the file of json
-     */
     public static void  loadFromFile() {
         cardList = cardDBAccess.readData(loggedInUser.getUserID());
     }
 
-    /**
-     * used to save the updated data in JSON file
-     * @param card cards given in CardMethod CardList
-     */
-    public static void saveCards(Card card) {
+    public static void saveCards(CardObject card) {
         cardDBAccess.saveData(loggedInUser.getUserID(), card);
     }
 
-    /**
-     * used to save the updated data in JSON file
-     * @param index the index card will be deleted in CardMethod CardList
-     */
     public static void saveDeleteCard(int index) {
         cardDBAccess.saveDeleteData(loggedInUser.getUserID(), index);
     }
 
-    public Card getCard(String cardID) {
+    public CardObject getCard(String cardID) {
         return cardDBAccess.readDataPoint(loggedInUser.getUserID(), cardID);
     }
 
@@ -66,122 +50,4 @@ public class CardController implements ControllerInterface {
         cardDBAccess.updateData(userID, cardID, amount);
     }
 
-    /**
-     * used to get the new if with no same id
-     */
-    @NotNull
-    public static String newId(String name) {
-        String id = "";
-        for (int i = 0; i < Math.min(name.length(), 3); i++) {
-            char c = name.replaceAll("\\s", "").charAt(i);
-            int k = c;
-            id += String.valueOf(k);
-        }
-        String insideId = id + getDifferentnumber(id);
-        return getIdForTest(insideId, id);
-    }
-
-    @NotNull
-    public static String getIdForTest(String insideId, String id) {
-        while (!checkId(insideId) | insideId.length() < 10) {
-            insideId = id + getDifferentnumber(id);
-        }
-        id = insideId;
-        return id;
-    }
-
-    /**
-     * used to check the repeating of id
-     * @param id random id that need to check if there is same id in file
-     */
-    public static boolean checkId(String id) {
-        loadFromFile();
-        for (Card card : cardList) {
-            if (Objects.equals(card.getId(), id)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * used to check the random part of id is with same index
-     * @param id random id given by the name
-     */
-    public static String getDifferentnumber(String id) {
-        long num = Math.round((Math.pow(10, (10 - id.length()))) * Math.random());
-        return threeCaseFor0(id, num);
-    }
-
-    @NotNull
-    public static String threeCaseFor0(String id, long num) {
-        switch (10 - id.length()) {
-            case 1:
-                return String.valueOf(num);
-            case 2:
-                if (num < 10) {
-                    return "0" + num;
-                }
-                else {
-                    return String.valueOf(num);
-                }
-            case 3:
-                if (num < 10) {
-                    return "00" + num;
-                }
-                else if (num < 100) {
-                    return "0" + num;
-                }
-                else {
-                    return String.valueOf(num);
-                }
-        }
-        return String.valueOf(num);
-    }
-
-    /**
-     * used to get the new update date for month and year
-     */
-    @NotNull
-    public static String newDate() {
-        LocalDate today = LocalDate.now();
-        int currentYear = today.getYear();
-        int currentMonth = today.getMonthValue();
-        return getDateForTest(currentMonth, currentYear);
-    }
-
-    @NotNull
-    public static String getDateForTest(int currentMonth, int currentYear) {
-        String month;
-        String expiryDate;
-        if (currentMonth < 10) {
-            month = "0" + currentMonth;
-        } else {
-            month = String.valueOf(currentMonth);
-        }
-        expiryDate = month + "/" + (currentYear + 5);
-        return expiryDate;
-    }
-
-    /**
-     * used to get the new code which is random with same index
-     */
-    @NotNull
-    public static String newCode() {
-        long num = Math.round(1000 * Math.random());
-        return getNewCodeForTest(num);
-    }
-
-    @NotNull
-    public static String getNewCodeForTest(long num) {
-        String securityCode;
-        if (num >= 100) {
-            securityCode = String.valueOf(num);
-        } else if (num >= 10) {
-            securityCode = "0" + num;
-        } else {
-            securityCode = "00" + num;
-        }
-        return securityCode;
-    }
 }
